Final authorization updates

+++ app/controllers/comments_controller.rb
@@ -2,15 +2,15 @@ class CommentsController < ApplicationController
   respond_to :html, :js
   
   def create
-    @topic = Topic.find(params[:topic_id])
-    @post = @topic.posts.find(params[:post_id])
+    @topic = Topic.find( params[:topic_id] )
+    @post = @topic.posts.find( params[:post_id] )
     @comments = @post.comments
 
-    @comment = current_user.comments.build(params[:comment])
+    @comment = current_user.comments.build( comment_params )
     @comment.post = @post
     @new_comment = Comment.new
     
-    authorize! :create, @comment, message: "You need be signed in to do that."
+    authorize @comment
 
     if @comment.save
       flash[:notice] = "Comment was created."
@@ -28,7 +28,7 @@ class CommentsController < ApplicationController
     @post = @topic.posts.find(params[:post_id])
 
     @comment = @post.comments.find(params[:id])
-    authorize! :destroy, @comment, message: "You need to own the comment to delete it."
+    authorize @comment
 
     if @comment.destroy
       flash[:notice] = "Comment was removed."
@@ -39,6 +39,16 @@ class CommentsController < ApplicationController
     respond_with(@comment) do |f|
       f.html { redirect_to [@topic, @post] }
     end
+  end
+
+  private
+
+  def comment_params
+    params.require(:comment).permit(
+      :body,
+      :topic_id,
+      :post_id
+    )
   end
 
 end
