+++ app/controllers/favorites_controller.rb
@@ -3,7 +3,7 @@ class FavoritesController < ApplicationController
     @topic = Topic.find(params[:topic_id])
     @post = @topic.posts.find(params[:post_id])
 
-    authorize! :create, Favorite, message: "You cannot do that"
+    authorize Favorite.new
 
     favorite = current_user.favorites.create(post: @post)
     
@@ -21,7 +21,7 @@ class FavoritesController < ApplicationController
     @post = @topic.posts.find(params[:post_id])
     @favorite = current_user.favorites.find(params[:id])
 
-    authorize! :destroy, @favorite, message: "You cannot do that."
+    authorize @favorite
 
     if @favorite.destroy
       flash[:notice] = "Removed favorite."
